fix: return OK when no value within /get-hashchain
When no value is found within /get-hashchain, there is still a valid non-membership-proof response. In order not to confuse clients that may think a non-success code means a failed request, we switch to a success code here. Clients have to parse the hashchain property to see if there is a valid value.