feat(tree): Return membership-proof when getting tree values (#136)
* feat(tree): Add struct for MembershipProof

* feat(tree): Return membership proof when getting tree values

Before this, only a non-membership-proof was returned when the value could not be found in the tree. Now, a proof is returned alongside the value when it can be found.

* feat(sequencer): Adapt to new tree get signature

* feat(webserver): Return membership-proof for /get-hashchain

* fix: return OK when no value within /get-hashchain

When no value is found within /get-hashchain, there is still a valid non-membership-proof response. In order not to confuse clients that may think a non-success code means a failed request, we switch to a success code here. Clients have to parse the hashchain property to see if there is a valid value.

* fix: return 500 instead of 400 for internal error

* docs: Add docs for HashchainResponse enum

* refac: Bring HashchainResponse enum members into scope

That way if-let constructs for ensuring some HashchainResponse value are way more readable

* refac: Pull out root variable

* test(tree): Verify membership proof when testing get