feat(common) Allow adding signed data (#141)
Introduces a new operation type, AddData, allowing users to add arbitrary data to accounts. The added data can optionally be signed as well.

* feat: Add operation to add signed data

* feat: Process signed data op in hashchain

* test: Method to add signed data to a TestAccount

* feat: tree delegates new op to hashchain

* refac: use usize for key indices

* feat: Added data op is always signed with key from chain

When solely relying on an external signature, this leaves prism vulnerable to replay attacks. As a result, ops to add data are always signed with a valid chain key and can optionally have an additional signature for the data to be added.

* feat: Process data op in prover

* chore: Incorporate new op code in zkvm elf