fix: reject genesis transactions with incorrect key or id
Without that fix, genesis transactions can be signed with a key different to the one being added. Also, the equality of operation and transaction ids is can be checked here. Currently, it is checked in snarkable_tree, but we have all required information for the check in this Account method too - so adding it here.