fix(common): reject incorrectly signed genesis transactions (#245)
* fix(common): reject genesis transactions with incorrect key or id

Without that fix, genesis transactions can be signed with a key different to the one being added. Also, the equality of operation and transaction ids is can be checked here. Currently, it is checked in snarkable_tree, but we have all required information for the check in this Account method too - so adding it here.

* test(common): Add tests for transaction processing in Accounts

There are similar tests in the tree crate, but these are testing tree specific functions (e.g. proof generation). Introducing a file for tests restricted to the scope of the common crate here (e.g. tests that do not care about database state, or proving).

* chore: Incorporate changes in zkvm elf