refactor: move sub-component config into node configs
Goal is to remove the burden to create EventChannel and DA/DB/whatever from devs. With this we provide 1-arg factory functions to create a node. Creation of sub-components is handled in the node factory function by invoking the sub-component's factory functions. Constructors to inject mock objects for testing remain in place.